Add a Browser layer for BDD tests
