+++ src/EzSystems/BehatBundle/Features/Context/FeatureContext.php
@@ -11,15 +11,12 @@
 
 namespace EzSystems\BehatBundle\Features\Context;
 
-use Behat\Behat\Context\Step;
 use Behat\Behat\Event\OutlineExampleEvent;
 use Behat\Behat\Event\ScenarioEvent;
 use Behat\MinkExtension\Context\MinkContext;
-use Behat\Mink\Exception\UnsupportedDriverActionException as MinkUnsupportedDriverActionException;
 use Behat\Symfony2Extension\Context\KernelAwareInterface;
 use eZ\Publish\Core\MVC\Symfony\SiteAccess;
 use PHPUnit_Framework_Assert as Assertion;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpKernel\KernelInterface;
 
 /**
@@ -98,80 +95,6 @@ class FeatureContext extends MinkContext implements KernelAwareInterface
     }
 
     /**
-     * @When /^I search for "([^"]*)"$/
-     */
-    public function iSearchFor( $searchPhrase )
-    {
-        $session = $this->getSession();
-        $searchField = $session->getPage()->findById( 'site-wide-search-field' );
-
-        Assertion::assertNotNull( $searchField, 'Search field not found.' );
-
-        $searchField->setValue( $searchPhrase );
-
-        // Ideally, using keyPress(), but doesn't work since no keypress handler exists
-        // http://sahi.co.in/forums/discussion/2717/keypress-in-java/p1
-        //     $searchField->keyPress( 13 );
-        //
-        // Using JS instead:
-        // Note:
-        //     $session->executeScript( "$('#site-wide-search').submit();" );
-        // Gives:
-        //     error:_call($('#site-wide-search').submit();)
-        //     SyntaxError: missing ) after argument list
-        //     Sahi.ex@http://<hostname>/_s_/spr/concat.js:3480
-        //     @http://<hostname>/_s_/spr/concat.js:3267
-        // Solution: Encapsulating code in a closure.
-        // @todo submit support where recently added to MinkCoreDriver, should us it when the drivers we use support it
-        try
-        {
-            $session->executeScript( "(function(){ $('#site-wide-search').submit(); })()" );
-        }
-        catch ( MinkUnsupportedDriverActionException $e )
-        {
-            // For drivers not able to do javascript we assume we can click the hidden button
-            $searchField->getParent()->findButton( 'SearchButton' )->click();
-        }
-
-        // Store for reuse in result page
-        $this->priorSearchPhrase = $searchPhrase;
-    }
-
-    /**
-     * @Then /^I am on the "([^"]*)"$/
-     */
-    public function iAmOnThe( $pageIdentifier )
-    {
-        $currentUrl = $this->getUrlWithoutQueryString( $this->getSession()->getCurrentUrl() );
-
-        $expectedUrl = $this->locatePath( $this->getPathByPageIdentifier( $pageIdentifier ) );
-
-        Assertion::assertEquals(
-            $expectedUrl,
-            $currentUrl,
-            "Unexpected URL of the current site. Expected: '$expectedUrl'. Actual: '$currentUrl'."
-        );
-    }
-
-    /**
-     * @Then /^(?:|I )want dump of (?:|the )page$/
-     */
-    public function iWantDumpOfThePage()
-    {
-        echo $this->getSession()->getPage()->getContent();
-    }
-
-    /**
-     * @When /^I go to the "([^"]*)"$/
-     */
-    public function iGoToThe( $pageIdentifier )
-    {
-        return array(
-            new Step\When( 'I am on "' . $this->getPathByPageIdentifier( $pageIdentifier ) . '"' ),
-        );
-    }
-
-    /**
      * Returns the path associated with $pageIdentifier
      *
      * @param string $pageIdentifier
@@ -203,74 +126,5 @@ class FeatureContext extends MinkContext implements KernelAwareInterface
         }
 
         return $url;
-    }
-
-    /**
-     * @Given /^I see search (\d+) result$/
-     */
-    public function iSeeSearchResults( $arg1 )
-    {
-        $resultCountElement = $this->getSession()->getPage()->find( 'css', 'div.feedback' );
-
-        Assertion::assertNotNull(
-            $resultCountElement,
-            'Could not find result count text element.'
-        );
-
-        Assertion::assertEquals(
-            "Search for \"{$this->priorSearchPhrase}\" returned {$arg1} matches",
-            $resultCountElement->getText()
-        );
-    }
-
-    /**
-     * @Given /^I am logged in as "([^"]*)" with password "([^"]*)"$/
-     */
-    public function iAmLoggedInAsWithPassword( $user, $password )
-    {
-        return array(
-            new Step\Given( 'I am on "/user/login"' ),
-            new Step\When( 'I fill in "Username" with "' . $user . '"' ),
-            new Step\When( 'I fill in "Password" with "' . $password . '"' ),
-            new Step\When( 'I press "Login"' ),
-            new Step\Then( 'I should be redirected to "/"' ),
-        );
-    }
-
-    /**
-     * @Then /^I should be redirected to "([^"]*)"$/
-     */
-    public function iShouldBeRedirectedTo( $redirectTarget )
-    {
-        $redirectForm = $this->getSession()->getPage()->find( 'css', 'form[name="Redirect"]' );
-
-        Assertion::assertNotNull(
-            $redirectForm,
-            'Missing redirect form.'
-        );
-
-        Assertion::assertEquals( $redirectTarget, $redirectForm->getAttribute( 'action' ) );
-    }
-
-    /**
-     * @Then /^I see (\d+) "([^"]*)" elements listed$/
-     */
-    public function iSeeListedElements( $count, $objectType )
-    {
-        $objectListTable = $this->getSession()->getPage()->find(
-            'xpath',
-            '//table[../h1 = "' . $objectType  . ' list"]'
-        );
-
-        Assertion::assertNotNull(
-            $objectListTable,
-            'Could not find listing table for ' . $objectType
-        );
-
-        Assertion::assertCount(
-            $count + 1,
-            $objectListTable->findAll( 'css', 'tr' ),
-            'Found incorrect number of table rows.'
-        );
     }
 }
